Fix compiler warnings
The compiler shows several warnings which should be addressed:

fileio.c:5: Pointer to the local statically allocated char array cannot
be returned from the function, because once the program leaves the scope
of the function, there is no guarantee that the array will be preserved.
Making the variable static is the easiest fix of this issue.

fileio/input_fitsWCS.c:93: The k variable can contain a signed integer
number (from -2147483646 to 2147483646) that will occupy between 3 and
11 bytes when formatted into a %03d string.  Note that the %03d format
specifies the *minimum* number of digits, but it does not prevent from
including number greater than 999.  The snprintf() function then does
the truncation, and thus prevents buffer overflow, but this is generally
quite a bad practice.  In this scenario the -Wformat-truncation compiler
argument generates the warning if a developer does not at least check if
the return value of snprintf() is not negative.

fileio/input_fitsWCS:97: This is one of the most dangerous pieces of
code within the ROBOSPECT that is most likely responsible for
unpredictable behavior.  The sprintf() function simply cannot read and
write into the same string at the same time.  Furthermore, space (' ')
as a flag of the format string do not have any function in "%s" format
and it is simply ignored by sprintf().  I have rewritten this code using
pointer arithmetic to avoid frequent NULL pointer checking.

models/continuum_blackbody.c:76: The undeclared chi variable contains a
random value during the first log_comment() call.  Although it is not
critical, accessing undeclared variable should be avoided in any
circumstances.  The iterations starts with chi square value 99e99 and
this is exactly what we should be printed during the first log_comment()
call.