Fix missing memory allocation error handling
The return value of the malloc() and realloc() calls was never checked
for an error.  Although this did not produce any warning, it is
generally a very bad practice.

My implementation of the memory allocation error handling might be
simplistic, but it represents the bare minimum that should have been
implemented a long ago.